requirements.txt
praw==7.7.1
python-dotenv==1.0.0
requests==2.32.5
Pillow==10.4.0
openai==1.35.0
